+++ openstack/autoscaling/v1/groups_hcs/list.go
@@ -1,0 +1,21 @@
+package groups_hcs
+
+import (
+	"github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
+)
+
+func List(client *golangsdk.ServiceClient, ops ListOptsBuilder) pagination.Pager {
+	url := client.ServiceURL("scaling_group")
+	if ops != nil {
+		q, err := ops.ToGroupListQuery()
+		if err != nil {
+			return pagination.Pager{Err: err}
+		}
+		url += q
+	}
+
+	return pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
+		return GroupPage{pagination.SinglePageBase(r)}
+	})
+}
