+++ openstack/autoscaling/v1/groups_hcs/requests.go
@@ -2,7 +2,6 @@ package groups_hcs
 
 import (
 	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
 )
 
 // CreateGroupBuilder is an interface from which can build the request of creating group
@@ -49,31 +48,6 @@ func (opts CreateOpts) ToGroupCreateMap() (map[string]interface{}, error) {
 	return golangsdk.BuildRequestBody(opts, "")
 }
 
-// CreateGroup is a method of creating group
-func Create(client *golangsdk.ServiceClient, opts CreateOptsBuilder) (r CreateResult) {
-	b, err := opts.ToGroupCreateMap()
-	if err != nil {
-		r.Err = err
-		return
-	}
-	_, r.Err = client.Post(createURL(client), b, &r.Body, &golangsdk.RequestOpts{
-		OkCodes: []int{200},
-	})
-	return
-}
-
-// DeleteGroup is a method of deleting a group by group id
-func Delete(client *golangsdk.ServiceClient, id string) (r DeleteResult) {
-	_, r.Err = client.Delete(deleteURL(client, id), nil)
-	return
-}
-
-// GetGroup is a method of getting the detailed information of the group by id
-func Get(client *golangsdk.ServiceClient, id string) (r GetResult) {
-	_, r.Err = client.Get(getURL(client, id), &r.Body, nil)
-	return
-}
-
 type ListOptsBuilder interface {
 	ToGroupListQuery() (string, error)
 }
@@ -91,21 +65,6 @@ func (opts ListOpts) ToGroupListQuery() (string, error) {
 		return "", err
 	}
 	return q.String(), err
-}
-
-func List(client *golangsdk.ServiceClient, ops ListOptsBuilder) pagination.Pager {
-	url := listURL(client)
-	if ops != nil {
-		q, err := ops.ToGroupListQuery()
-		if err != nil {
-			return pagination.Pager{Err: err}
-		}
-		url += q
-	}
-
-	return pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
-		return GroupPage{pagination.SinglePageBase(r)}
-	})
 }
 
 // UpdateOptsBuilder is an interface which can build the map paramter of update function
@@ -137,21 +96,6 @@ func (opts UpdateOpts) ToGroupUpdateMap() (map[string]interface{}, error) {
 	return golangsdk.BuildRequestBody(opts, "")
 }
 
-// Update is a method which can be able to update the group via accessing to the
-// autoscaling service with Put method and parameters
-func Update(client *golangsdk.ServiceClient, id string, opts UpdateOptsBuilder) (r UpdateResult) {
-	body, err := opts.ToGroupUpdateMap()
-	if err != nil {
-		r.Err = err
-		return
-	}
-
-	_, r.Err = client.Put(updateURL(client, id), body, &r.Body, &golangsdk.RequestOpts{
-		OkCodes: []int{200},
-	})
-	return
-}
-
 type ActionOptsBuilder interface {
 	ToActionMap() (map[string]interface{}, error)
 }
@@ -170,24 +114,8 @@ func doAction(client *golangsdk.ServiceClient, id string, opts ActionOptsBuilder
 		r.Err = err
 		return
 	}
-	_, r.Err = client.Post(enableURL(client, id), &b, nil, &golangsdk.RequestOpts{
+	_, r.Err = client.Post(client.ServiceURL("scaling_group", id, "action"), &b, nil, &golangsdk.RequestOpts{
 		OkCodes: []int{204},
 	})
 	return
-}
-
-// Enable is an operation by which can make the group enable service
-func Enable(client *golangsdk.ServiceClient, id string) (r ActionResult) {
-	opts := ActionOpts{
-		Action: "resume",
-	}
-	return doAction(client, id, opts)
-}
-
-// Disable is an operation by which can be able to pause the group
-func Disable(client *golangsdk.ServiceClient, id string) (r ActionResult) {
-	opts := ActionOpts{
-		Action: "pause",
-	}
-	return doAction(client, id, opts)
 }
