--- openstack/autoscaling/v1/groups_hcs/urls.go
@@ -1,35 +1,0 @@
-package groups_hcs
-
-import (
-	"log"
-
-	"github.com/opentelekomcloud/gophertelekomcloud"
-)
-
-const resourcePath = "scaling_group"
-
-func createURL(c *golangsdk.ServiceClient) string {
-	ur := c.ServiceURL(resourcePath)
-	log.Printf("[DEBUG] Create URL is: %#v", ur)
-	return ur
-}
-
-func deleteURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id)
-}
-
-func getURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id)
-}
-
-func listURL(c *golangsdk.ServiceClient) string {
-	return c.ServiceURL(resourcePath)
-}
-
-func enableURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id, "action")
-}
-
-func updateURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id)
-}
